@@ -6,7 +6,7 @@ xfns_default : list(callable)
     The default VAEP features.
 
 """
-from typing import List, Optional
+from typing import Optional
 
 import socceraction.atomic.spadl as spadlcfg
 from socceraction.vaep.base import VAEP
@@ -72,7 +72,7 @@ class AtomicVAEP(VAEP):
 
     def __init__(
         self,
-        xfns: Optional[List[fs.FeatureTransfomer]] = None,
+        xfns: Optional[list[fs.FeatureTransfomer]] = None,
         nb_prev_actions: int = 3,
     ) -> None:
         xfns = xfns_default if xfns is None else xfns
