@@ -1,6 +1,6 @@
 """Implements the feature tranformers of the VAEP framework."""
 from functools import wraps
-from typing import Any, Callable, List, Union, no_type_check
+from typing import Any, Callable, Union, no_type_check
 
 import numpy as np  # type: ignore
 import pandas as pd  # type: ignore
@@ -12,12 +12,12 @@ from socceraction.spadl.schema import SPADLSchema
 
 SPADLActions = DataFrame[SPADLSchema]
 Actions = Union[DataFrame[SPADLSchema], DataFrame[AtomicSPADLSchema]]
-GameStates = List[Actions]
+GameStates = list[Actions]
 Features = DataFrame[Any]
 FeatureTransfomer = Callable[[GameStates], Features]
 
 
-def feature_column_names(fs: List[FeatureTransfomer], nb_prev_actions: int = 3) -> List[str]:
+def feature_column_names(fs: list[FeatureTransfomer], nb_prev_actions: int = 3) -> list[str]:
     """Return the names of the features generated by a list of transformers.
 
     Parameters
@@ -139,7 +139,7 @@ def simple(actionfn) -> FeatureTransfomer:
     """
 
     @wraps(actionfn)
-    def _wrapper(gamestates: List[Actions]) -> pd.DataFrame:
+    def _wrapper(gamestates: list[Actions]) -> pd.DataFrame:
         if not isinstance(gamestates, (list,)):
             gamestates = [gamestates]
         X = []
